+++ order/src/main/java/com/southwind/repository/OrderRepository.java
@@ -8,8 +8,8 @@ public interface OrderRepository {
     public void save(Order order);  //传入的是order
     public List<Order> findAll(int index, int limit);
                                                         // session中保存的登录用户
-    public List<Order> findAllByUid(int index, int limit, int uid);//某用户的点单
+    public List<Order> findAllByUid(int index, int limit, long uid);//某用户的点单
 
-    public int countByUid(int uid);
+    public int countByUid(long uid);
 
 }
